+++ datacount.js
@@ -1,0 +1,3 @@
+export function datacount(value) {
+  return value;
+}
