+++ server.js
@@ -1,6 +1,6 @@
-// Initialize the JS client
-
 // Make a request
-export default function show() {
-  console.log("splitted file");
-}
+var SUPABASE_URL = "https://csnsvhmzohqmmlvjexpu.supabase.co";
+var SUPABASE_KEY =
+  "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImNzbnN2aG16b2hxbW1sdmpleHB1Iiwicm9sZSI6ImFub24iLCJpYXQiOjE2NDY2NDI1MTQsImV4cCI6MTk2MjIxODUxNH0.RaTg59kmRNNRQlfwFUhuOIzjbZujClUketfNBHRxEW8";
+
+export default supabase.createClient(SUPABASE_URL, SUPABASE_KEY);
